refactor(x834-context): remove separate localDateTime variable to track time. Now just use a single dateTime and pass it around as needed.
