+++ src/main/java/com/fastChickensHR/edi/common/segments/DTPSegment.java
@@ -10,7 +10,7 @@ package com.fastChickensHR.edi.common.segments;
 import com.fastChickensHR.edi.common.dates.DateFormat;
 import com.fastChickensHR.edi.common.dates.DateFormatter;
 import com.fastChickensHR.edi.common.exception.ValidationException;
-import com.fastChickensHR.edi.x834.common.x834Context;
+import com.fastChickensHR.edi.x834.x834Context;
 import lombok.Getter;
 
 import java.time.LocalDateTime;
