+++ src/main/java/com/fastChickensHR/edi/x834/header/FunctionalGroupHeader.java
@@ -10,7 +10,7 @@ package com.fastChickensHR.edi.x834.header;
 import com.fastChickensHR.edi.common.segments.Segment;
 import com.fastChickensHR.edi.common.dates.DateFormatter;
 import com.fastChickensHR.edi.common.exception.ValidationException;
-import com.fastChickensHR.edi.x834.common.x834Context;
+import com.fastChickensHR.edi.x834.x834Context;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.experimental.Accessors;
