+++ src/main/java/com/fastChickensHR/edi/x834/loop2000/DependentMember.java
@@ -9,7 +9,7 @@ package com.fastChickensHR.edi.x834.loop2000;
 
 import com.fastChickensHR.edi.common.segments.Segment;
 import com.fastChickensHR.edi.common.exception.ValidationException;
-import com.fastChickensHR.edi.x834.common.x834Context;
+import com.fastChickensHR.edi.x834.x834Context;
 import lombok.Getter;
 import lombok.Setter;
