+++ src/main/java/com/fastChickensHR/edi/x834/loop2000/loop3000/HealthCoverageDates.java
@@ -9,7 +9,7 @@ package com.fastChickensHR.edi.x834.loop2000.loop3000;
 
 import com.fastChickensHR.edi.common.segments.DTPSegment;
 import com.fastChickensHR.edi.common.exception.ValidationException;
-import com.fastChickensHR.edi.x834.common.x834Context;
+import com.fastChickensHR.edi.x834.x834Context;
 import lombok.Getter;
 
 /**
