+++ src/test/java/com/fastChickensHR/edi/common/segments/DTPSegmentTest.java
@@ -9,7 +9,7 @@ package com.fastChickensHR.edi.common.segments;
 
 import com.fastChickensHR.edi.common.dates.DateFormat;
 import com.fastChickensHR.edi.common.exception.ValidationException;
-import com.fastChickensHR.edi.x834.common.x834Context;
+import com.fastChickensHR.edi.x834.x834Context;
 import org.junit.jupiter.api.Test;
 
 import java.time.LocalDateTime;
