+++ src/test/java/com/fastChickensHR/edi/x834/loop1000B/PayerTest.java
@@ -8,7 +8,7 @@
 package com.fastChickensHR.edi.x834.loop1000B;
 
 import com.fastChickensHR.edi.common.exception.ValidationException;
-import com.fastChickensHR.edi.x834.common.x834Context;
+import com.fastChickensHR.edi.x834.x834Context;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
