+++ src/test/java/com/fastChickensHR/edi/x834/loop2000/MemberLevelDetailTest.java
@@ -8,7 +8,7 @@
 package com.fastChickensHR.edi.x834.loop2000;
 
 import com.fastChickensHR.edi.common.exception.ValidationException;
-import com.fastChickensHR.edi.x834.common.x834Context;
+import com.fastChickensHR.edi.x834.x834Context;
 import com.fastChickensHR.edi.x834.loop2000.data.ConfidentialityCode;
 import com.fastChickensHR.edi.x834.loop2000.data.EmploymentStatusCode;
 import com.fastChickensHR.edi.x834.loop2000.data.MaintenanceReasonCode;
